requirements.txt
typing-extensions~=4.14.1
urllib3~=2.5.0
appium-python-client~=4.3.0
Pillow==10.4.0
pytesseract==0.3.10
geopy~=2.4.1
setuptools~=80.9.0
gpxpy~=1.6.2
adb_pywrapper~=1.2.0
requests~=2.32.3
